refact: uint32 => int

+++ quantum.go
@@ -10,57 +10,57 @@ var R2 float32 = 0.70710678118
 
 type QuantumGate struct {
 	gate   string
-	qubit  uint32
-	target uint32
+	qubit  int
+	target int
 	angle  float32
 }
 
 type QuantumCircuit struct {
-	number_qubits uint32
+	number_qubits int
 	circuit       []QuantumGate
 }
 
-func NewQuantumCircuit(q uint32) QuantumCircuit {
+func NewQuantumCircuit(q int) QuantumCircuit {
 	return QuantumCircuit{number_qubits: q, circuit: []QuantumGate{}}
 }
 
-func (qc *QuantumCircuit) AddGate(gate string, qubit, target uint32, angle float32) {
+func (qc *QuantumCircuit) AddGate(gate string, qubit, target int, angle float32) {
 	element := QuantumGate{gate, qubit, target, angle}
 	qc.circuit = append(qc.circuit, element)
 }
 
-func (qc *QuantumCircuit) X(qubit uint32) {
+func (qc *QuantumCircuit) X(qubit int) {
 	qc.AddGate("X", qubit, 0, 0.0)
 }
 
-func (qc *QuantumCircuit) Y(qubit uint32) {
+func (qc *QuantumCircuit) Y(qubit int) {
 	qc.RZ(qubit, math.Pi)
 	qc.X(qubit)
 }
 
-func (qc *QuantumCircuit) Z(qubit uint32) {
+func (qc *QuantumCircuit) Z(qubit int) {
 	qc.RZ(qubit, math.Pi)
 }
 
-func (qc *QuantumCircuit) H(qubit uint32) {
+func (qc *QuantumCircuit) H(qubit int) {
 	qc.AddGate("H", qubit, 0, 0.0)
 }
 
-func (qc *QuantumCircuit) CX(control, target uint32) {
+func (qc *QuantumCircuit) CX(control, target int) {
 	qc.AddGate("CX", control, target, 0.0)
 }
 
-func (qc *QuantumCircuit) RX(qubit uint32, angle float32) {
+func (qc *QuantumCircuit) RX(qubit int, angle float32) {
 	qc.AddGate("RX", qubit, 0, angle)
 }
 
-func (qc *QuantumCircuit) RZ(qubit uint32, angle float32) {
+func (qc *QuantumCircuit) RZ(qubit int, angle float32) {
 	qc.H(qubit)
 	qc.RX(qubit, angle)
 	qc.H(qubit)
 }
 
-func (qc *QuantumCircuit) RY(qubit uint32, angle float32) {
+func (qc *QuantumCircuit) RY(qubit int, angle float32) {
 	qc.RX(qubit, math.Pi/2.0)
 	qc.H(qubit)
 	qc.RX(qubit, angle)
@@ -70,8 +70,8 @@ func (qc *QuantumCircuit) RY(qubit uint32, angle float32) {
 
 type QuantumSimulator struct {
 	circuit       []QuantumGate
-	number_qubits uint32
-	bits          uint32
+	number_qubits int
+	bits          int
 	state_vector  []complex128
 }
 
@@ -112,14 +112,14 @@ func (self *QuantumSimulator) Run() {
 	for _, quantum_gate := range self.circuit {
 		if quantum_gate.gate == "X" || quantum_gate.gate == "H" || quantum_gate.gate == "RX" {
 			// Don't punish me typecasting gods
-			for counter_qubit := uint32(0); counter_qubit <= uint32(math.Pow(2, float64(quantum_gate.qubit)))-1; counter_qubit++ {
-				for counter_state := uint32(0); counter_state <= uint32(math.Pow(2, float64(self.number_qubits-quantum_gate.qubit-1)))-1; counter_state++ {
-					qb0 := counter_qubit + uint32(math.Pow(2, float64(quantum_gate.qubit))+1)*counter_state
-					qb1 := qb0 + uint32(math.Pow(2, float64(quantum_gate.qubit)))
+			for counter_qubit := 0; counter_qubit <= int(math.Pow(2, float64(quantum_gate.qubit)))-1; counter_qubit++ {
+				for counter_state := 0; counter_state <= int(math.Pow(2, float64(self.number_qubits-quantum_gate.qubit-1)))-1; counter_state++ {
+					qb0 := counter_qubit + int(math.Pow(2, float64(quantum_gate.qubit))+1)*counter_state
+					qb1 := qb0 + int(math.Pow(2, float64(quantum_gate.qubit)))
 
 					if quantum_gate.gate == "X" {
 						swapF := reflect.Swapper(self.state_vector)
-						swapF(int(qb0), int(qb1))
+						swapF(qb0, qb1)
 					}
 					if quantum_gate.gate == "H" {
 						a, b := superposition(self.state_vector[qb0], self.state_vector[qb1])
@@ -137,18 +137,18 @@ func (self *QuantumSimulator) Run() {
 			low, high := lohi(quantum_gate.qubit, quantum_gate.target)
 			swapF := reflect.Swapper(self.state_vector)
 			// 0..2**low
-			for cx0 := uint32(0); cx0 <= uint32(math.Pow(2, float64(low))); cx0++ {
-			  // 2**(high-low-1)
+			for cx0 := 0; cx0 <= int(math.Pow(2, float64(low))); cx0++ {
+				// 2**(high-low-1)
 				limit := math.Pow(2, float64(high-low-1))
 				// 0..limit
-				for cx1 := uint32(0); cx1 <= uint32(limit); cx1++ {
-				  // 0..2**self.number_qubits - high - 1
-					for cx2 := uint32(0); cx2 <= uint32(math.Pow(2, float64(self.number_qubits-high-1))); cx2++ {
-					  // cx0 + 2**low+1 * cx1 + 2**high+1 * cx2 + 2**quantum_gate.qubit
-						qb0 := cx0 + uint32(math.Pow(2, float64(low+1)))*cx1 + uint32(math.Pow(2, float64(high+1)))*cx2 + uint32(math.Pow(2, float64(quantum_gate.qubit)))
+				for cx1 := 0; cx1 <= int(limit); cx1++ {
+					// 0..2**self.number_qubits - high - 1
+					for cx2 := 0; cx2 <= int(math.Pow(2, float64(self.number_qubits-high-1))); cx2++ {
+						// cx0 + 2**low+1 * cx1 + 2**high+1 * cx2 + 2**quantum_gate.qubit
+						qb0 := cx0 + int(math.Pow(2, float64(low+1)))*cx1 + int(math.Pow(2, float64(high+1)))*cx2 + int(math.Pow(2, float64(quantum_gate.qubit)))
 						// qb0 + 2**quantum_gate.target
-						qb1 := qb0 + uint32(math.Pow(2, float64(quantum_gate.target)))
-						swapF(int(qb0), int(qb1))
+						qb1 := qb0 + int(math.Pow(2, float64(quantum_gate.target)))
+						swapF(qb0, qb1)
 					}
 				}
 			}
@@ -163,7 +163,7 @@ func (self *QuantumSimulator) print() {
 	}
 }
 
-func lohi(q0, q1 uint32) (uint32, uint32) {
+func lohi(q0, q1 int) (int, int) {
 	if q0 < q1 {
 		return q0, q1
 	} else {
@@ -174,7 +174,7 @@ func lohi(q0, q1 uint32) (uint32, uint32) {
 func main() {
 	qc := NewQuantumCircuit(15)
 	qc.H(0)
-	for qubit := uint32(1); qubit <= 15; qubit++ {
+	for qubit := 1; qubit <= 15; qubit++ {
 		qc.H(qubit)
 		qc.CX(qubit-1, qubit)
 	}
